requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
PyMySQL==1.1.0
requests==2.31.0
urllib3==2.1.0
colorama==0.4.6
